Workaround for the peek function
This would solve the problem of recursing deep into some types of objects. 
In the case of the mongoose object we had the problem with, it's a moot point anyway because even when the copy is a deep copy, any modification in the copy will modify the original (if saved to the DB, of course) because the _id of the object gets copied anyway. But this has nothing to do with the copy itself.